feat: moving configs and hooks to separate files in layout to clean up the code
